requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-split-settings==1.2.0
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
djangorestframework==3.14.0
